Conflicts:
	kexi/core/KoContextHelp.cpp
	kexi/doc/dev/CHANGELOG-Kexi-js

+++ krita/ui/kis_canvas_resource_provider.h
@@ -112,8 +112,8 @@ public:
     void setMirrorVertical(bool mirrorVertical);
     bool mirrorVertical() const;
 
-    void setOpacity(int opacity);
-    int opacity();
+    void setOpacity(qreal opacity);
+    qreal opacity();
 
 public slots:
